feat(net): zero-cost sock-state when not on linux
