+++ compio-net/src/socket/mod.rs
@@ -2,8 +2,6 @@ use std::{
     future::Future,
     io,
     mem::{ManuallyDrop, MaybeUninit},
-    ops::Deref,
-    sync::atomic::Ordering,
 };
 
 use compio_buf::{
@@ -25,10 +23,7 @@ use compio_driver::{
 use compio_runtime::{Attacher, Runtime, fd::PollFd};
 use futures_util::{Stream, StreamExt, future::Either};
 use socket2::{Domain, Protocol, SockAddr, Socket as Socket2, Type};
-#[cfg(feature = "sync")]
-use synchrony::sync::atomic::AtomicI8;
-#[cfg(not(feature = "sync"))]
-use synchrony::unsync::atomic::AtomicI8;
+use sys::SocketState;
 
 use crate::Incoming;
 
@@ -47,65 +42,27 @@ cfg_if::cfg_if! {
     }
 }
 
-// We are not on the IO_URING driver and hence retrieving socket state is
-// not supported.
-const UNSUPPORTED: i8 = -1;
+cfg_if::cfg_if! {
+    if #[cfg(target_os = "linux")] {
+        #[path = "linux.rs"]
+        mod sys;
+    } else {
+        mod sys {
+            #[derive(Default, Clone, Debug)]
+            pub(super) struct SocketState;
 
-// The socket was empty after the receive operation.
-const EMPTY: i8 = 0;
+            impl SocketState {
+                pub(super) fn new() -> Self {
+                    SocketState
+                }
 
-// The socket was not-empty after the last receive operation and has more
-// data to be read.
-const NON_EMPTY: i8 = 1;
+                pub(super) fn get(&self) -> Option<bool> {
+                    None
+                }
 
-#[derive(Debug)]
-struct SocketState {
-    state: AtomicI8,
-}
-
-impl Default for SocketState {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-impl SocketState {
-    fn new() -> Self {
-        Self {
-            state: AtomicI8::new(-1),
+                pub(super) fn set(&self, _: &compio_driver::Extra) {}
+            }
         }
-    }
-
-    fn get(&self) -> Option<bool> {
-        match self.load(Ordering::Relaxed) {
-            UNSUPPORTED => None,
-            EMPTY => Some(false),
-            NON_EMPTY => Some(true),
-            _ => unreachable!(),
-        }
-    }
-
-    fn set(&self, extra: &compio_driver::Extra) {
-        if let Ok(n) = extra.sock_nonempty() {
-            self.store(n as i8, Ordering::Relaxed);
-        }
-    }
-}
-
-impl Clone for SocketState {
-    fn clone(&self) -> Self {
-        let current = self.state.load(Ordering::Relaxed);
-        Self {
-            state: AtomicI8::new(current),
-        }
-    }
-}
-
-impl Deref for SocketState {
-    type Target = AtomicI8;
-
-    fn deref(&self) -> &Self::Target {
-        &self.state
     }
 }
